Some improvements on component and directory structure
