+++ components/atoms/avatar/__tests__/stories.js
@@ -9,9 +9,12 @@ import React from 'react'
 import { Avatar } from '../'
 import Readme from '../README.md'
 
+import { number, withKnobs } from '@storybook/addon-knobs'
+
 export default {
   title: 'Atoms/Avatar',
   component: Avatar,
+  decorators: [withKnobs],
   parameters: {
     readme: {
       sidebar: Readme
@@ -19,7 +22,7 @@ export default {
   }
 }
 
-export const defaultStory = () => <Avatar content='Avatar' />
+export const defaultStory = () => <Avatar content='Avatar' size={number('Size', 60)} />
 
 defaultStory.story = {
   name: 'Default'
