+++ components/atoms/avatar/variables.js
@@ -1,0 +1,9 @@
+/**
+ * Theme - Variables - Avatar
+ */
+import { COLOUR } from '../../theme/variables/colour'
+
+export const AVATAR = {
+  background: COLOUR.primary,
+  colour: COLOUR.black
+}
