+++ components/theme/variables/index.js
@@ -8,6 +8,7 @@ import { TYPOGRAPHY } from './typography'
 // Components
 import { FOOTER } from '../../organisms/footer/variables'
 import { GRID } from '../../atoms/grid/variables'
+import { AVATAR } from '../../atoms/avatar/variables'
 import { NAVBAR } from '../../molecules/navbar/variables'
 import { SIDEBAR } from '../../molecules/sidebar/variables'
 import { TABS } from '../../molecules/tabs/variables'
@@ -16,6 +17,7 @@ export const Theme = {
   COLOUR,
   FOOTER,
   GRID,
+  AVATAR,
   NAVBAR,
   SIDEBAR,
   TABS,
